Moved the LevelUp component to the Add Entry component, to be proc'd when the Add Entry modal is closed, hopefully conserving processing power
