+++ src/components/App/App.jsx
@@ -22,8 +22,8 @@ import StatsPage from "../PageComponents/StatsPage/StatsPage";
 import AchievementsPage from "../PageComponents/AchievementsPage/AchievementsPage";
 import AvatarsPage from "../PageComponents/Avatars/AvatarsPage/AvatarsPage"
 
-import AddLogButton from "../ButtonComponents/AddLogButton/AddLogButton";
-import LevelUpModal from "../LevelUpPopup/LevelUp.jsx";
+import AddLogButton from "../ButtonComponents/AddLogComponent/AddLogButton.jsx";
+import LevelUpModal from "../ButtonComponents/AddLogComponent/LevelUpPopup/LevelUp.jsx";
 
 //importing to wrap app for the date picker
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
@@ -133,11 +133,6 @@ export default function App() {
           <>
             {// only want the log activity button to show if the user is logged in
               user.id ? <AddLogButton /> : <></>
-            }
-          </>
-          <>
-            {
-              user.id ? <LevelUpModal /> : <></>
             }
           </>
           <Footer />
